Move views to separate modules
