+++ console/views/organization.py
@@ -1,18 +1,6 @@
-"""console.views"""
+"""console.views.organizations"""
 from django.shortcuts import render
 from django.views import View
-
-
-class DashboardView(View):
-    """
-    Dashboard view
-    """
-
-    def get(self, request):
-        """
-        Render the dashboard.html
-        """
-        return render(request, 'console/dashboard.html')
 
 
 class OrganizationsView(View):
